VC: Improve halting GUI: Minor improvements Controller: Handle timing in here rather than in VC

+++ src/Instruction.java
@@ -54,7 +54,7 @@ public class Instruction {
         mnemonicLookup.put("HLT", Mnemonics.HLT);
     }
 
-    public boolean execute(MainMemory mm, byte PC, Register[] registers) {
+    public boolean execute(MainMemory mm, int PC, Register[] registers) {
         switch (opcode) {
             case Mnemonics.NOP:
                 break;
